break up agent hierarchy to reflect what is Basic agent behavior and what is NLIP behavior
