+++ nlip_agents/agents/basic_agent.py
@@ -34,28 +34,8 @@ MODEL = "anthropic/claude-3-7-sonnet-20250219"
 # PROMPTS
 #
 
-INSTRUCTION = """
+TOOLS_INSTRUCTION = """
 You are an agent with tools.  When calling a tool, make sure to match the type signature of the tool.
-"""
-
-NLIP = """
-You are an NLIP Agent.
-
-NLIP is an acronym for "Natural Language Interaction Protocol."  The NLIP project aims to define a protocol for the following use cases.
-- Agent to Agent interactions in Natural Language
-- User-Agent to Agent protocol
-
-An NLIP Agent, when defined, is given a system instruction that describes its unique capabilities.
-
-One of the first requests an NLIP Agent will be asked to fulfill is to describe its NLIP Capabilities.
-When you are asked to describe your NLIP Capabilities, you should respond with a response of the format:
-    AGENT:NAME
-    CAPABILITY1:description, CAPABILITY2:description, CAPABILITY3:description, ...
-- where NAME is your name
-- CAPABILITITY1, CAPABILITY2 and CAPABILITY3 are dictionary keys.  The description associated with each should be unique.
-- it is important that you include your NAME
-- it is important that the capabilities are described in a means that another LLM can understand them
-
 """
 
 class BasicAgent:
@@ -85,11 +65,7 @@ class BasicAgent:
         self.messages: list[Any] = [
             {
                 "role": "system",
-                "content": INSTRUCTION
-            },
-            {
-                "role": "system",
-                "content": NLIP
+                "content": TOOLS_INSTRUCTION
             }
         ]
