+++ nlip_agents/agents/nlip_agent.py
@@ -1,94 +1,37 @@
 #
-# An NLIP Coordinator Agent has tools for communicating with other NLIP Agents.
-# An additional system instruction helps refine its behavior.
-
-import logging
+# An NLIP Agent knows how to respond to the request for NLIP Capabilities
+#
 
 import asyncio
+import logging
+from typing import Callable
+
 from .basic_agent import BasicAgent
+from .basic_agent import MODEL
 
-from typing import Any
-import httpx
-import json
 
-# Import NLIP
-from nlip_agents.http_client.nlip_async_client import NlipAsyncClient
-from nlip_sdk.nlip import NLIP_Factory, NLIP_Message
-from urllib.parse import urlparse
-
-from pydantic import AnyHttpUrl
-
-# map host->session
-sessions = { }
-
-# Use the NLIP logger in this package
 logger = logging.getLogger("NLIP")
 
-# MODEL = 'llama3.2:latest'
-# MODEL = 'ollama_chat/llama3.2:latest'
-# MODEL = 'ollama_chat/llama3-groq-tool-use:
-MODEL = "anthropic/claude-3-7-sonnet-20250219"
+NLIP_INSTRUCTION = """
+You are an NLIP Agent.
 
+NLIP is an acronym for "Natural Language Interaction Protocol."  The NLIP project aims to define a protocol for the following use cases.
+- Agent to Agent interactions in Natural Language
+- User-Agent to Agent protocol
 
-#
-# Make a connection and return a status string
-#
-    
-async def connect_to_server(url: AnyHttpUrl):
-    """Connect to the server and return a message"""
-    try:
-        parsed_url = urlparse(url)
-        scheme = parsed_url.scheme
-        netloc = parsed_url.netloc
-    except Exception as e:
-        instance.text = f"Exception: {e}"
-        return
+An NLIP Agent, when defined, is given a system instruction that describes its unique capabilities.
 
-    # Establish the URL and return a connection message
-    await asyncio.sleep(1.0)
-    client = NlipAsyncClient.create_from_url(f"{scheme}://{netloc}/nlip/")
+One of the first requests an NLIP Agent will be asked to fulfill is to describe its NLIP Capabilities.
+When you are asked to describe your NLIP Capabilities, you should respond with a response of the format:
+    AGENT:NAME
+    CAPABILITY1:description, CAPABILITY2:description, CAPABILITY3:description, ...
+- where NAME is your name
+- CAPABILITITY1, CAPABILITY2 and CAPABILITY3 are dictionary keys.  The description associated with each should be unique.
+- it is important that you include your NAME
+- it is important that the capabilities are described in a means that another LLM can understand them
 
-    # Remember this client for this server
-    hashkey = f"{scheme}://{netloc}"
-    sessions[hashkey] = client
-
-    logger.info(f"Saved {netloc} with client {client}")
-    return f"Connected to {scheme}://{netloc}/"
-
-#
-# Send a message to a named host and get a response
-#
-
-# async def send_to_server(url: AnyHttpUrl, msg: str) -> NLIP_Message:
-async def send_to_server(url: AnyHttpUrl, msg: str) -> dict:
-    parsed_url = urlparse(url)
-    scheme = parsed_url.scheme
-    netloc = parsed_url.netloc
-
-    # Look up the client for this server
-    hashkey = f"{scheme}://{netloc}" # netloc includes host and port, if specified
-    client = sessions[hashkey]
-
-    nlip_message = NLIP_Factory.create_text(msg)
-    logger.info(f"Sending message: {msg}")
-    nlip_resp = await client.async_send(nlip_message)
-    logger.info(f"Received: {nlip_resp.model_dump()}")
-    # return nlip_resp.model_dump() # did not work
-    # return nlip_resp.extract_text() # worked
-    return str(nlip_resp.model_dump())
-
-
-NLIP = """
-You are an advanced NLIP Agent with the capability to speak to other NLIP Agents.
-You have two tools for this purpose:
-- connect_to_server
-- send_to_server
-
-When you are asked to connect to a server at a specific URL, use the connect_to_server tool with that URL to establish a connection.
-If the response to that tool begins with: "Connected to ", then the connection is valid.  Otherwise, it is not.
-For a valid connection, you should follow the connect_to_server tool call with a tool call of send_to_server to the same URL with the string: "What are your NLIP Capabilities?"
-The remote Agent will respond with its [NAME] and capabilities.  Take note of this information, especially the NAME.  In future requests, if a user asks for you to send a request to NAME you should use the send_to_server tool with the URL that was associated with NAME and use the request as the msg: argument.
 """
+
 
 class NlipAgent(BasicAgent):
 
@@ -96,19 +39,22 @@ class NlipAgent(BasicAgent):
                  name: str,
                  model: str = MODEL,
                  instruction: str = None,
-                 tools = [connect_to_server, send_to_server]
+                 tools: list[Callable] = [ ]
                  ):
 
-        super().__init__(name, model=model, tools=tools)
+        super().__init__(name, model, NLIP_INSTRUCTION, tools)
 
-        self.add_instruction("You are an agent with tools for querying other NLIP Agent Servers")
-        self.add_instruction(NLIP)
-
+        # any additional instructions
         if instruction:
             self.add_instruction(instruction)
-    
 
-# Test program for running with stdin
+
+
+
+#
+# Test program for trying out a simple agent with the console
+#
+
 async def main():
     logger.info(f"WELCOME")
     if len(sys.argv) < 0:
@@ -116,7 +62,7 @@ async def main():
         sys.exit(1)
 
     agent = NlipAgent(
-        "Margaret"
+        name="Bart"
     )
     
     try:
